# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GRIDIX Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "List Models",
        "description": "Catalogue models, each flagged with whether a node is serving it right now.\n\nAvailability is reported rather than filtered: a developer whose model went dark\nshould see that it exists and is offline, not watch it vanish from the list.",
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/chat": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Public Chat",
        "description": "Stream a free chat completion. No account, no balance, no ledger entry.\n\nStreamed rather than unary because it is what makes the queue tolerable: a caller who\nwaited for a slot starts seeing tokens the moment they get one.",
        "operationId": "public_chat_public_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/images": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Public Image",
        "description": "Generate a free image \u2014 CLOSED, and closed by the safety system rather than a flag.\n\nTwo independent reasons this cannot serve anything today, and the check order matters:\n\n1. NO MODERATION IS CONFIGURED. `image_generation_available()` is false whenever the\n   moderator cannot make decisions, and the default moderator cannot. Public,\n   unauthenticated image generation without CSAM and NCII screening is not something to\n   ship behind a TODO, so the door is shut by the absence of the control rather than by\n   a boolean someone could flip without noticing what it guards.\n2. NOTHING CAN GENERATE AN IMAGE. The node package serves chat only and answers\n   `images.generations` with 501; no node advertises an image model.\n\nThe quota below it is real and tested regardless, so that enabling this route later is\nsupplying a moderator \u2014 not also discovering that the limit was never written.",
        "operationId": "public_image_public_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Public Image Public Images Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "description": "One turn. OpenAI-shaped so existing clients work unchanged."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "nodes": {
            "type": "integer",
            "title": "Nodes"
          },
          "input_usdc_per_mtok": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Input Usdc Per Mtok"
          },
          "output_usdc_per_mtok": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Output Usdc Per Mtok"
          },
          "usdc_per_image": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Usdc Per Image"
          },
          "context_window": {
            "type": "integer",
            "title": "Context Window"
          }
        },
        "type": "object",
        "required": [
          "id",
          "modality",
          "available",
          "nodes",
          "input_usdc_per_mtok",
          "output_usdc_per_mtok",
          "usdc_per_image",
          "context_window"
        ],
        "title": "ModelInfo",
        "description": "A catalogue model and whether the network is serving it right now."
      },
      "ModelsResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelsResponse",
        "description": "Everything GRIDIX can serve, with prices."
      },
...
      "PublicChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 1,
            "title": "Messages"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 2048.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "default": 512
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "PublicChatRequest",
        "description": "A free chat turn.\n\n``model`` is accepted but not trusted: whatever arrives, the served model is the free\none. Taking the caller's value would make the allowlist a suggestion."
      },
      "PublicImageRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "PublicImageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}